// src/features/daily-groove/lib/grooves.generated.ts
export const GROOVES: Groove[] = [
  {
    id: 'groove-01',
    audioSrc: '/grooves/groove-01.mp3',
    name: 'Rusted Shuffle',
    bpm: 105,
    scale: 'C mixolydian',
    chord: 'C7',
    progression: 'C7–Em7♭5–B♭maj7–Fmaj7',
    root: 'C',
    flavour: 'Mixolydian',
    bars: 4,
  },
  {
    id: 'groove-02',
    audioSrc: '/grooves/groove-02.mp3',
    name: 'Dusty Ravine',
    bpm: 96,
    scale: 'E dorian',
    chord: 'Em7',
    progression: 'Em7–Bm7–C♯m7♭5',
    root: 'E',
    flavour: 'Dorian',
    bars: 4,
  },
  {
    id: 'groove-03',
    audioSrc: '/grooves/groove-03.mp3',
    name: 'Feathered Pocket',
    bpm: 103,
    scale: 'E♭ dorian',
    chord: 'E♭m7',
    progression: 'E♭m7–A♭7–Fm7',
    root: 'E♭',
    flavour: 'Dorian',
    bars: 4,
  },
  {
    id: 'groove-04',
    audioSrc: '/grooves/groove-04.mp3',
    name: 'Quiet Drizzle',
    bpm: 103,
    scale: 'E mixolydian',
    chord: 'E7',
    progression: 'E7–Amaj7–Bm7–Amaj7',
    root: 'E',
    flavour: 'Mixolydian',
    bars: 4,
  },
  {
    id: 'groove-05',
    audioSrc: '/grooves/groove-05.mp3',
    name: 'Marbled Orbit',
    bpm: 90,
    scale: 'E♭ blues',
    chord: 'E♭7',
    progression: 'E♭7–B♭7–A♭7',
    root: 'E♭',
    flavour: 'Blues',
    bars: 4,
  },
  {
    id: 'groove-06',
    audioSrc: '/grooves/groove-06.mp3',
    name: 'Wandering Parlour',
    bpm: 83,
    scale: 'A blues',
    chord: 'A7',
    progression: 'A7–D7–E7–D7',
    root: 'A',
    flavour: 'Blues',
    bars: 4,
  },
  {
    id: 'groove-07',
    audioSrc: '/grooves/groove-07.mp3',
    name: 'Humid Lagoon',
    bpm: 91,
    scale: 'G minor',
    chord: 'Gm7',
    progression: 'Gm7–B♭maj7–Cm7',
    root: 'G',
    flavour: 'Minor',
    bars: 4,
  },
  {
    id: 'groove-08',
    audioSrc: '/grooves/groove-08.mp3',
    name: 'Gilded Lagoon',
    bpm: 91,
    scale: 'F♯ minor',
    chord: 'F♯m7',
    progression: 'F♯m7–Amaj7–A♭m7♭5',
    root: 'F♯',
    flavour: 'Minor',
    bars: 4,
  },
  {
    id: 'groove-09',
    audioSrc: '/grooves/groove-09.mp3',
    name: 'Hazy Awning',
    bpm: 126,
    scale: 'C♯ lydian',
    chord: 'C♯maj7',
    progression: 'C♯maj7–E♭7–Fm7–B♭m7',
    root: 'C♯',
    flavour: 'Lydian',
    bars: 4,
  },
  {
    id: 'groove-10',
    audioSrc: '/grooves/groove-10.mp3',
    name: 'Humid Alley',
    bpm: 121,
    scale: 'F lydian',
    chord: 'Fmaj7',
    progression: 'Fmaj7–Am7–G7–Dm7',
    root: 'F',
    flavour: 'Lydian',
    bars: 4,
  },
  {
    id: 'groove-11',
    audioSrc: '/grooves/groove-11.mp3',
    name: 'Midnight Parlour',
    bpm: 120,
    scale: 'B major',
    chord: 'Bmaj7',
    progression: 'Bmaj7–Emaj7–C♯m7',
    root: 'B',
    flavour: 'Major',
    bars: 4,
  },
  {
    id: 'groove-12',
    audioSrc: '/grooves/groove-12.mp3',
    name: 'Wandering Basement',
    bpm: 130,
    scale: 'A major',
    chord: 'Amaj7',
    progression: 'Amaj7–A♭m7♭5–F♯m7',
    root: 'A',
    flavour: 'Major',
    bars: 4,
  },
  {
    id: 'groove-13',
    audioSrc: '/grooves/groove-13.mp3',
    name: 'Sunken Lullaby',
    bpm: 79,
    scale: 'A♭ phrygian',
    chord: 'A♭m7',
    progression: 'A♭m7–Amaj7–F♯m7–Amaj7',
    root: 'A♭',
    flavour: 'Phrygian',
    bars: 4,
  },
  {
    id: 'groove-14',
    audioSrc: '/grooves/groove-14.mp3',
    name: 'Smoky Harbour',
    bpm: 72,
    scale: 'D phrygian',
    chord: 'Dm7',
    progression: 'Dm7–Gm7–E♭maj7',
    root: 'D',
    flavour: 'Phrygian',
    bars: 4,
  },
  {
    id: 'groove-15',
    audioSrc: '/grooves/groove-15.mp3',
    name: 'Plush Alley',
    bpm: 80,
    scale: 'A♭ harmonic minor',
    chord: 'A♭mMaj7',
    progression: 'A♭mMaj7–Gdim7–C♯m7–Gdim7',
    root: 'A♭',
    flavour: 'Harmonic minor',
    bars: 4,
  },
  {
    id: 'groove-16',
    audioSrc: '/grooves/groove-16.mp3',
    name: 'Salted Cassette',
    bpm: 69,
    scale: 'G harmonic minor',
    chord: 'GmMaj7',
    progression: 'GmMaj7–Am7♭5–F♯dim7',
    root: 'G',
    flavour: 'Harmonic minor',
    bars: 4,
  },
]